reduced audio clipping, apu still needs a LOT of work (#42)

+++ internal/gameboy/gameboy.go
@@ -83,7 +83,7 @@ func (g *GameBoy) SendCommand(command emulator.CommandPacket) emulator.ResponseP
 	return emulator.ResponsePacket{}
 }
 
-func (g *GameBoy) AttachAudioListener(player func([]byte)) {
+func (g *GameBoy) AttachAudioListener(player func([]float32)) {
 	g.APU.AttachPlayback(player)
 }
